LC-4317-Learning catalogue changes for course tag management

+++ src/main/resources/db/migration/mysql/V1.1.0__create-course-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE course
 (
-    id                 smallint unsigned AUTO_INCREMENT,
+    id                 BIGINT AUTO_INCREMENT,
     uid                varchar(255)            NOT NULL,
     title              varchar(255)            NOT NULL,
